+++ src/main/song_lullaby.h
@@ -2,11 +2,12 @@
 
 #include "song.h"
 #include "pitches.h"
+#include <stdint.h>
 
 #define LULLABY_SOUND_LENGTH 54
 #define LULLABY_TEMPO 1500
 
 // BRAM'S LULLABY
-extern int *lullabyMelody;
-extern int *lullabyDuration;
+extern short int *lullabyMelody;
+extern uint8_t *lullabyDuration;
 extern Song *lullaby;
